requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
Flask-WhooshAlchemy==0.56
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
Tempita==0.5.2
WTForms==2.0.2
Werkzeug==0.9.6
Whoosh==2.6.0
blinker==1.3
decorator==3.4.0
gunicorn==19.3.0
itsdangerous==0.24
pbr==0.10.7
six==1.9.0
sqlalchemy-migrate==0.9.4
sqlparse==0.1.14
